Separate functions for data generation and data insert
